xwax: Add a quadrature phase tracker
The current pitch filter works under the assumption that zero crossings
are never skipped. Unfortunately this happens quite frequently when
frequencies are very high during e.g. scratching or backspins. Especially
during backspins the pitch filter was being starved by always feeding a
quarter rotation displacement, when in reality the displacement was a half
rotation or three quarters even.

This commit fixes this issue by mapping the quadrants integers 1..4.
The last quadrant we were in, is compared to the current one and the
difference is calculated. A difference of 1 corresponds to a 90° phase angle.
This rotation angle is then quantized to timecode revolutions.

This fixes backspins and improves pitch robustness when the needle runs
over a scratch or during audio buffer dropouts.

fixup