xwax: Add ZERO_THRESHOLD to struct timecode_def
While working on the phase tracker I discovered that many zero crossings
are skipped. There were also too many direction changes detected for the
MK2s. I could improve this by lifting the threshold.

This commits also adds it to the timecode definition. It's not clear if
the default threshold works well for other timecodes then the Seratos.
With this structure thresholds can be changed per timecode.

We might also want to dynamically calibrate a threshold to the
RMS value of the carrier wave. This would probably yield better results
than using these fixed ones.