fix: Map the shiftbutton messages to decks 1/3 and 2/4
I previously assumed the shift button sent messages on a per-deck
basis. While learning to DJ on 4 decks simultanously, I noticed
that the shift button doesn't work on decks 3 and 4. It turns out
that regardless of the currently active deck on each side, the
shift button doesn't change the channel it originates from and
thus it would be interpreted as Decks 1/2 being shifted instead
of the currently active decks 3/4. This may also be a firmware bug.

The easiest fix is just to dispatch the shift button to both decks
on each side.