feat(components): JogwheelBasic: Disable scratching when leaving `vinylMode`
Since I don't think its common to disable vinylMode while having
your hand still on the platter and expecting to still scratch,
disable scratching as soon as `vinylMode` has been disabled.
This also helps in the case where a platter may be malfunctioning
(leaving a touch as registered even though you released it). In
that case, the track remains paused and there is nothing the user
can do about it. With this patch, they can workaround the issue
by disabling vinylMode, getting the deck unstuck.