fix: initialize `allshader::WaveformWidget::m_type`
Otherwise this would not only be a case of using unitialized memory,
but even if the memory was 0 by chance (seemed to be the case), the
waveform viewer would think the waveform is actually empty and thus
refuse any interaction with it. So this fixes both a memory safety
issue and a user-facing bug.

Fixes #15374