Fixed incorrect folder hierarchy importing
If a folder contained a node that was something else that a folder or a playlist (SMARTLIST for example) then the node would not be added to the current_path but it would be removed from current_path when the node closed. This would cause the folder hierarchy to get out of sync and display incorrectly.

This now processes every node but only adds playlists and non-empty folder to the item tree.