Update Numark-Party-Mix.scripts.js
So in the original version of this everything worked except the Jog wheels. They worked when you select, "Scratch" on the DJ deck, but they did not work when you did not have scratch selected. I found a script in a forum long ago by DarkPoubelle https://mixxx.discourse.group/t/numark-party-mix-midi-mapping/16712/8 who fixed the wheel issue however, some other features did not work. I figured out a way to combine the logic from DarkPoubelle's script with this one, and using this new script you can use the jog wheels both while the "Scratch" button is selected and is not selected. Very useful fork I would say, I have mentioned "DarkPoubelle" in the script as a credit to them fixing the issue first.