fix(ExternalLibraryFeature): Refactored createPlaylistModelForPlaylist()
This now receives a QVariant which is the data part of the TreeItem the request to create the playlist originates from. This allow each implementation of the LibraryFeature class to correctly retrieve the playlist path according to how it stored it in there in the first place.

This fixes all playlist importing from other programs like Rekordbox, Serato, etc.. which were silently failing before.