Performance: batch-import playlist items from Rhythmbox
This is the equivalent of running all insertions between a "BEGIN TRANSACTION" and a "COMMIT".
In my case (~20 playlists on a ~8000-item library), it turns the Rhythmbox import from a 4-minute process to a 10-second one.
As a side-effect, it also stops showing plenty of "Mixxx is not responding" dialogs during that process.